Improve bye message triggering
Say bye when page is unloaded. This is kind of tricky due to Safari not
supporting unload event correctly. For regular page navigation, pagehide
works best, but it's not triggered when tab is closed.

Page reload events:
  - beforeunload
  - pagehide
  - unload

Page navigation events:
  - beforeunload
  - pagehide (only triggered if user or script didn't cancel unload)

Tab close events:
  - beforeunload

In order to say bye (and possibly undo it) correctly we need to use both
beforeunload & pagehide.